requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pillow>=10.0.0
reportlab>=4.0.0
img2pdf>=0.5.0
python-multipart>=0.0.9
jinja2>=3.1.0
